Rename Homework1-RebasingDecimalInteger.m to RebasingDecimalInteger.m
Write a Matlab function that converts a non negative integer written in decimal base to its equivalent form expressed on an arbitrary base .
Store the output in a vector. On input, the function must accept the number n, the base b, and the maximum number of digits m. On output, the function must return a vector of length m. If the number causes an overflow (i.e., it is bigger than what can be represented with m digits on that base), the vector must be filled with NaNs. The program must check that b>1. If  it must return a vector of -1's. 
The return expression must contain the Most Significant Digit in the first entry of the vector. 
For example, if the number is 99=1*81+2*9, the base is 3 and m = 7, then the answer is 0 0 1 0 2 0 0.
For the same number, but the base in 5 and m=4 then the answer is 0 3 4 4.