Create pi4Error.m
It is well known that [pi/4 = arctan(1)] Use the Taylor series for the arctan(1) at the origin to write [pi/4] as a series. Write a function that returns the error when [pi/4] is evaluated, truncating the series after n terms. For the exact value of [pi] use the builtin matlab variable pi.