Create polyNewton_interpNewton.m
For this test, you need to write two functions: polyNewton(X,Y) and interpNewton(a,X,x). 
X and Y=f(X) are vectors which contains the values to be interpolated. The first function returns the coefficients of the Newton's polynomial
The second uses the output of the first and X to calculate the interpolated  values at points contained in x. 
The twist is that you are allowed one and only ONE for loop in each function AND you are not allowed 
to allocate any extra memory aside from a, which stores the output in polyNewton, and y which stores the output in interpNewton 
(the dummy index in the for loop does not count against memory allocation).